Add locale definition file parsing class.
